создает модуль данных
